+++ CHANGELOG.md
@@ -1,5 +1,30 @@
 # Changelog
 
+## [1.27.6] - 2026-02-11
+
+### Added
+- dual-runtime-sqlite
+
+## [1.27.5] - 2026-02-11
+
+### Added
+- dual-runtime-sqlite
+
+## [1.27.4] - 2026-02-11
+
+### Added
+- dual-runtime-sqlite
+
+## [1.27.3] - 2026-02-11
+
+### Added
+- current work
+
+## [1.27.2] - 2026-02-11
+
+### Added
+- current work
+
 ## [1.27.1] - 2026-02-11
 
 ### Removed